% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_to.R
\name{entrez_to_attribute}
\alias{entrez_to_attribute}
\title{Obtain gene annotations}
\usage{
entrez_to_attribute(df, species, attribute, dir_annotation = getwd(),
  verbose = TRUE)
}
\arguments{
\item{df}{A data.frame containing a column named "entrezgene_id".}

\item{species}{One of "mouse" or "human".}

\item{attribute}{A gene attribute. Examples include: "gene_biotype",
"description", "chromosome_name", etc. See listAttributes() for
more options.}

\item{dir_annotation}{The directory to store annotation reference.}

\item{verbose}{Set to FALSE to avoid messages.}
}
\value{
The data.frame df with an added column containing gene symbols.
}
\description{
Obtain gene annotations
}
