% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_to_GO.R
\name{entrez_to_GO}
\alias{entrez_to_GO}
\title{Obtain gene ontology (GO) annotations}
\usage{
entrez_to_GO(df, species)
}
\arguments{
\item{df}{A data.frame containing a column named "entrezgene_id".}

\item{species}{One of "mouse" or "human".}
}
\value{
A data.frame containing GO annotations for the genes in df.
}
\description{
Obtain gene ontology (GO) annotations
}
