% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reactome_pathways.R
\name{find_similar_pathways}
\alias{find_similar_pathways}
\title{Find pathways that overlap with a reference pathway.}
\usage{
find_similar_pathways(pathway_list, pathway, threshold = 0.9)
}
\arguments{
\item{pathway_list}{A list of pathways obtained from
\code{\link{get_reactome_pathways}}.}

\item{pathway}{A reference pathway.}

\item{threshold}{A percentage between 0 and 1. If two pathways
overlap by more than this amount, then they are considered similar.}
}
\value{
The indices to pathway_list that contain pathways similar
to the reference pathway.
}
\description{
Find pathways that overlap with a reference pathway.
}
