% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reactome_pathways.R
\name{get_reactome_pathways}
\alias{get_reactome_pathways}
\title{Obtain reactome pathways}
\usage{
get_reactome_pathways(species = c("human", "mouse"), file_name = NULL,
  dir_reactome = getwd())
}
\arguments{
\item{species}{One of "mouse" or "human".}

\item{file_name}{The file name for pathway data. If null, defaults to
"reactome_[species].rds".}

\item{dir_reactome}{The directory to store the pathway data. Defaults to
the current working directory.}
}
\value{
A named list of vectors. Each vector corresponds to a Reactome pathway
and contains the entrezgene IDs of the genes in that pathway.
}
\description{
Obtain reactome pathways
}
