% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mart.R
\name{init_mart_human}
\alias{init_mart_human}
\title{Initialize biomaRt for hsapiens}
\usage{
init_mart_human()
}
\value{
None. Creates a global variable called mart_human.
}
\description{
Initialize biomaRt for hsapiens
}
