% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mart.R
\name{init_mart_mouse}
\alias{init_mart_mouse}
\title{Initialize biomaRt for mmusculus}
\usage{
init_mart_mouse()
}
\value{
None. Creates a global variable called mart_mouse.
}
\description{
Initialize biomaRt for mmusculus
}
