% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reactome_pathways.R
\name{remove_redundant_pathways}
\alias{remove_redundant_pathways}
\title{Modify a pathway list to combine overlapping pathways.}
\usage{
remove_redundant_pathways(pathway_list, threshold = 0.9)
}
\arguments{
\item{pathway_list}{A list of pathways obtained from
\code{\link{get_reactome_pathways}}.}

\item{threshold}{A percentage between 0 and 1. If two pathways
overlap by more than this amount, they are combined into one pathway.}
}
\value{
A modified list with overlapping pathways combined together.
}
\description{
Modify a pathway list to combine overlapping pathways.
}
