% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_to_entrez.R
\name{symbol_to_entrez}
\alias{symbol_to_entrez}
\title{Obtain entrezgene IDs for gene symbols}
\usage{
symbol_to_entrez(df, dir_annotation = getwd(), verbose = TRUE)
}
\arguments{
\item{df}{A data.frame containing a column named either "hgnc_symbol" or
"mgi_symbol".}

\item{dir_annotation}{The directory to store annotation reference.}

\item{verbose}{Set to FALSE to avoid messages.}
}
\value{
The data.frame df with an added column containing entrezgene IDs.
}
\description{
Obtain entrezgene IDs for gene symbols
}
